requirements.txt
# 联通网盘统一服务依赖库
# 安装命令: pip install -r requirements.txt

# Web框架
Flask==2.3.3
Werkzeug==2.3.7

# 加密库
pycryptodome==3.19.0

# HTTP请求库
requests==2.31.0
urllib3==2.0.7

# 并发处理
gevent==23.9.1
greenlet==3.0.1

# 生产环境WSGI服务器
gunicorn==21.2.0
waitress==2.1.2

# 其他依赖
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4

# 可选：性能监控
psutil==5.9.6

# GUI打包依赖
pyinstaller>=5.0
